Silence a couple of warnings.

+++ StashAppDelegate.mm
@@ -661,7 +661,7 @@
 	[self updateUI];
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item
 {
 	if (item == nil)
 	{
@@ -685,7 +685,7 @@
     return [item expandable];
 }
 
--(int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+-(NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
     if (item == nil)
 	{
